update for arch

+++ setup.sh
@@ -21,11 +21,16 @@ case "$OSTYPE" in
             spotify
     ;;
     "linux"*)
-        echo "run on linux"
-        apt update -y
-        apt install vim git zsh curl tmux -y
+	if [ -e /etc/arch-release ]; then
+	    pacman -Syyu
+            pacman -Sy vim git zsh curl tmux -y
+	else
+	    echo "run on linux"
+            apt update -y
+            apt install vim git zsh curl tmux -y
+	fi
         sh -c "$(curl -fsSL https://starship.rs/install.sh)" -- -y
-        curl --proto =h'ttps' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y # install Rust
+        curl --proto ='https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y # install Rust
     ;;
     *)
         echo "can not detect the OSTYPE"
